+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2021-04-30
+### Added
+- Support for Python 3.9
+
+### Changed
+- Added tests for hook functionality, fixed minor error, removed WIP tag
+- Updated log statement for "Starting simulation ..."
+
 ## [0.6.0] - 2021-02-20
 ### Changed
 - Moved Ray and dependencies to optional extension `extension-backend-ray`